* warp changes:    * implemented shader variants that make use of a argument buffer that store all materials (images) -> this is enabled when a device supports argument buffers and images in argument buffers    * when using a material argument buffer, this is now only set once and then reused for all shader calls    * changed depth buffer device-side access to read-only    * fixed renderer ownership/retention in the present renderer

+++ common/camera/camera.cpp
@@ -151,7 +151,7 @@ void camera::run() {
 			}
 			else ignore_next_rotation--;
 			
-			const float2 center_point(float2(floor::get_width(), floor::get_height()) * 0.5f);
+			const float2 center_point(float2(float(floor::get_width()), float(floor::get_height())) * 0.5f);
 			SDL_WarpMouseInWindow(floor::get_window(),
 								  (int)roundf(center_point.x), (int)roundf(center_point.y));
 		}
